refactor(themes.ts): rename themes file
